split _core.py to subfiles
